Move Seeds into parser
